+++ src/main/java/com/mobble/mobbleserver/domain/article/Article.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.article;
 
 import com.mobble.mobbleserver.domain.club.core.Club;
-import com.mobble.mobbleserver.domain.common.entity.BaseEntity;
+import com.mobble.mobbleserver.domain.common.BaseEntity;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
