+++ src/main/java/com/mobble/mobbleserver/domain/article/ArticleContent.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.article;
 
 import com.mobble.mobbleserver.domain.article.error.ArticleError;
-import com.mobble.mobbleserver.domain.common.exception.DomainException;
+import com.mobble.mobbleserver.domain.exception.DomainException;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
