+++ src/main/java/com/mobble/mobbleserver/domain/chat/message/ChatMessage.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.chat.message;
 
 import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
-import com.mobble.mobbleserver.domain.common.entity.CreatedAtEntity;
+import com.mobble.mobbleserver.domain.common.CreatedAtEntity;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
