+++ src/main/java/com/mobble/mobbleserver/domain/club/core/Club.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.domain.club.core;
 
 import com.mobble.mobbleserver.domain.ClubCategory.ClubCategory;
 import com.mobble.mobbleserver.domain.address.Address;
-import com.mobble.mobbleserver.domain.common.entity.BaseEntity;
+import com.mobble.mobbleserver.domain.common.BaseEntity;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.club.ClubErrorCode;
 import jakarta.persistence.*;
