+++ src/main/java/com/mobble/mobbleserver/domain/clubMember/ClubMember.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.domain.clubMember;
 
 import com.mobble.mobbleserver.domain.article.ArticleType;
 import com.mobble.mobbleserver.domain.club.core.Club;
-import com.mobble.mobbleserver.domain.common.entity.BaseEntity;
+import com.mobble.mobbleserver.domain.common.BaseEntity;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
